Moved device selector helper
