+++ client/src/components/DeviceSelectorHelper.ts
